restify-back/requirements.txt
django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
django-rest-swagger==2.2.0
drf-yasj==1.20.0